+++ src/TypeInspectors/GetNamespace.php
@@ -62,7 +62,7 @@ class GetNamespace
      *         - if the string does not contain the name of a defined
      *           class / interface / trait
      */
-    public function __invoke($item)
+    public function getNamespace($item)
     {
         return self::from($item);
     }